+++ tests/unit/browsers/test_browser_pool.py
@@ -107,7 +107,7 @@ async def test_with_default_plugin_constructor(server_url: URL) -> None:
 async def test_new_page_with_existing_id() -> None:
     async with BrowserPool() as browser_pool:
         page_1 = await browser_pool.new_page()
-        with pytest.raises(ValueError, match='Page with ID: .* already exists.'):
+        with pytest.raises(ValueError, match=r'Page with ID: .* already exists.'):
             await browser_pool.new_page(page_id=page_1.id)
 
 
@@ -115,7 +115,7 @@ async def test_new_page_with_invalid_plugin() -> None:
     plugin_1 = PlaywrightBrowserPlugin(browser_type='chromium')
     plugin_2 = PlaywrightBrowserPlugin(browser_type='firefox')
     async with BrowserPool([plugin_1]) as browser_pool:
-        with pytest.raises(ValueError, match='Provided browser_plugin is not one of the plugins used by BrowserPool.'):
+        with pytest.raises(ValueError, match=r'Provided browser_plugin is not one of the plugins used by BrowserPool.'):
             await browser_pool.new_page(browser_plugin=plugin_2)
 
 
@@ -139,13 +139,13 @@ async def test_methods_raise_error_when_not_active() -> None:
 
     assert browser_pool.active is False
 
-    with pytest.raises(RuntimeError, match='BrowserPool is not active.'):
+    with pytest.raises(RuntimeError, match=r'BrowserPool is not active.'):
         await browser_pool.new_page()
 
-    with pytest.raises(RuntimeError, match='BrowserPool is not active.'):
+    with pytest.raises(RuntimeError, match=r'BrowserPool is not active.'):
         await browser_pool.new_page_with_each_plugin()
 
-    with pytest.raises(RuntimeError, match='BrowserPool is already active.'):
+    with pytest.raises(RuntimeError, match=r'BrowserPool is already active.'):
         async with browser_pool, browser_pool:
             pass